Added classes and display lines for round trip time display.
Display lines: Smoothed-Round-Trip-Time, Retransmit-Delay.
Classes: Config, NamedInt replaced by NamedLong, NsAsMsNamedLong.

+++ DagBrowser7/src/allClasses/DefaultLongLike.java
@@ -6,7 +6,7 @@ public class DefaultLongLike
   {
 	  private long theL;
 
-	  public DefaultLongLike( int theL ) // Constructor. 
+	  public DefaultLongLike( long theL ) // Constructor. 
 		  {
 		  	this.theL= theL;
         }
